+++ packages/ruleset-bundler/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @stoplight/spectral-ruleset-bundler [1.6.3](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-ruleset-bundler-1.6.2...@stoplight/spectral-ruleset-bundler-1.6.3) (2025-04-23)
+
+
+### Bug Fixes
+
+* **functions:** export or function ([#2812](https://github.com/stoplightio/spectral/issues/2812)) ([03532a5](https://github.com/stoplightio/spectral/commit/03532a5193e52094eb1a11c7ed8337b353be47c8))
+
 ## @stoplight/spectral-ruleset-bundler [1.6.2](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-ruleset-bundler-1.6.1...@stoplight/spectral-ruleset-bundler-1.6.2) (2025-03-07)